moves echo/envoycontrol test dependencies to separate packages
