+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/config/envoy/ResponseWithBody.kt
@@ -1,7 +1,7 @@
 package pl.allegro.tech.servicemesh.envoycontrol.config.envoy
 
 import okhttp3.Response
-import pl.allegro.tech.servicemesh.envoycontrol.config.containers.EchoContainer
+import pl.allegro.tech.servicemesh.envoycontrol.config.echo.EchoContainer
 
 data class ResponseWithBody(val response: Response, val body: String) {
     fun isFrom(echoContainer: EchoContainer) = body.contains(echoContainer.response)
